website/backend/requirements.txt
wall-library
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
